requirements.txt
Pillow==9.5.0
Pillow==11.2.1
streamlit==1.41.1
~treamlit==1.23.1
